mpa account create with default amont, pizzaservice account, and tran…

+++ src/main/java/whz/pti/eva/praktikum_03/boundary/OrderedController.java
@@ -13,10 +13,12 @@ import whz.pti.eva.praktikum_03.common.CurrentUserUtil;
 import whz.pti.eva.praktikum_03.domain.Ordered;
 import whz.pti.eva.praktikum_03.dto.CartDTO;
 import whz.pti.eva.praktikum_03.dto.CustomerDTO;
+import whz.pti.eva.praktikum_03.dto.PayActionResponseDTO;
 import whz.pti.eva.praktikum_03.enums.Role;
 import whz.pti.eva.praktikum_03.security.domain.CurrentUser;
 import whz.pti.eva.praktikum_03.service.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -34,6 +36,7 @@ public class OrderedController {
     private CustomerService customerService;
     private OrderedService orderedService;
     private OrderedItemService orderedItemService;
+    private SmmpService smmpService;
 
     private static final Logger log = LoggerFactory.getLogger(ItemController.class);
 
@@ -48,13 +51,14 @@ public class OrderedController {
      * @param orderedItemService the ordered item service
      */
     @Autowired
-    public OrderedController(PizzaService pizzaService, CartService cartService, ItemService itemService, CustomerService customerService, OrderedService orderedService, OrderedItemService orderedItemService) {
+    public OrderedController(PizzaService pizzaService, CartService cartService, ItemService itemService, CustomerService customerService, OrderedService orderedService, OrderedItemService orderedItemService, SmmpService smmpService) {
         this.pizzaService = pizzaService;
         this.cartService = cartService;
         this.itemService = itemService;
         this.customerService = customerService;
         this.orderedService = orderedService;
         this.orderedItemService = orderedItemService;
+        this.smmpService = smmpService;
     }
 
     /**
@@ -72,9 +76,25 @@ public class OrderedController {
             CustomerDTO currentCustomer = customerService.findByUserId(currentUser.getUser().getId());
             CartDTO customersCart = cartService.findCartByCustomer(currentCustomer.getId());
             if (customersCart.getQuantity() != 0) {
-                log.info("Creating ordered and ordered items. Delete items from customers cart");
-                Ordered ordered = orderedItemService.addOrderedItem(customersCart, currentCustomer);
-                itemService.deleteItems(customersCart, currentCustomer);
+                PayActionResponseDTO payActionResponseDTO = smmpService.doPayAction(currentCustomer.getLoginName(), "ps", "get");
+                BigDecimal bigDecimal = new BigDecimal(payActionResponseDTO.getDescription());
+                BigDecimal bigDecimal1 = cartService.calculateTotalPriceOfPizzaInItemsInCart(customersCart);
+                if (bigDecimal.compareTo(bigDecimal1) >= 0 ) {
+                    log.info("Balance is "+ bigDecimal);
+                    log.info("Ordered is "+ bigDecimal1);
+                    smmpService.doPayAction(currentCustomer.getLoginName(), "ps", "transfer ps "+bigDecimal1);
+                    log.info("Creating ordered and ordered items. Delete items from customers cart");
+                    Ordered ordered = orderedItemService.addOrderedItem(customersCart, currentCustomer);
+                    itemService.deleteItems(customersCart, currentCustomer);
+                }
+                else {
+                    log.info("Balance is less");
+                }
+
+
+
+
+
             }
             List<Ordered> orderedList = orderedService.findAllByCustomerId(currentCustomer.getId());
             model.addAttribute("orderedList", orderedList);
